clean updated

+++ clean.sh
@@ -1,2 +1,6 @@
-#! /bin/bash
-rm -r ./bin/*
+#!/bin/bash
+if [ "$(ls -A ./bin)" ]; then
+  rm -r ./bin/*
+else
+  echo "The bin directory is already empty."
+fi
